requirements.txt
Django>=2.0.1
mysqlclient
djangorestframework
django-cors-headers
google-cloud-pubsub==1.3.1